[WIP] DO NOT REVIEW cmd/compile: add basic block counters
Current patch adds the counters to the AST and SSA nodes. The counters
are loaded from the pprof file, no profile format changes needed. Currently
implemented:

 + Loading counters to AST nodes by line number information
 + The AST counter propagation algorithm
 + Moving counters to SSA nodes from AST nodes
 + Adding profile usage in basic block layout pass (just
   proof-of-concept)
 + Print counters to the html dumps

The patch shows how we can load basic block counters from the pprof
file, how we can propagate them, and how to use them. The patch was
tested on the go1 benchmark set, and it showed that depending on
bb-layout pass the test fannkuch can be faster from 3.5% to 12.2%
on the intel core machine (the xeon machine does not show this improvement).

To use basic block counters you should add an option -bbpgo:

go test -a -bbpgo -pgo=go1.prof

This patch just shows the main idea of basic block counters loading.
Before we can use it in the compiler, the following steps should be done:

 * Improve the counter propagation on the AST and SSA. Currently there
   are some propagations, but they still may be improved
 * Improve basic block layout and register allocation passes with
   profile information

Fixes #65466

Change-Id: I5d6be7d87f384625259a9ba794744a652060de4e